+++ composer.json
@@ -285,6 +285,11 @@
       "ignore-duplicates": true
     }
   },
+  "autoload-dev": {
+    "psr-4": {
+      "Drupal\\Tests\\hs_bugherd\\": "docroot/modules/humsci/hs_bugherd/tests"
+    }
+  },
   "scripts": {
     "post-drupal-scaffold-cmd": [
       "cd docroot && patch -p1 <../patches/htaccess-saml.patch",
